Improved command line interface.
Added [] overloaded operator for accessing ArgumentParser option values.

Implemented analysis program option and added checks for lack of arguments.

Added arguments for specifying seperate audio and database directories. Implemented automatic setting of database directory as the same as the audio directory if no value is provided on the command line.